More

+++ 3D Molecular Computation Package/src/simulation/TimeUpdatable.java
@@ -1,5 +1,7 @@
 package simulation;
 
+import java.math.BigDecimal;
+
 public interface TimeUpdatable extends Updatable{
-	public void update(double milliseconds);
+	public void update(BigDecimal milliseconds);
 }
